+++ src/elc/florian/tool/Utils.java
@@ -1,9 +1,9 @@
 package elc.florian.tool;
 
 public class Utils {
-    public static Boolean sorted(float[] l) {
-        float a = l[0];
-        for (float b : l) {
+    public static Boolean sorted(double[] l) {
+        double a = l[0];
+        for (double b : l) {
             if (!(a <= b)) {
                 return false;
             }
