+++ src/elc/florian/tool/listener/MyKeyListener.java
@@ -13,7 +13,7 @@ public class MyKeyListener implements KeyListener {
     public void keyPressed(KeyEvent e) {
         if (KeyEvent.getKeyText(e.getKeyCode()).equals("R")) {
             long end_time = System.currentTimeMillis();
-            float s = (float) (end_time - Main.start_time) /1000;
+            double s = (double) (end_time - Main.start_time) /1000;
 
             System.out.println(Main.refresh + " frames depuis " + s + " secondes");
             System.out.println(Main.refresh / s + " frames/s");
@@ -24,32 +24,32 @@ public class MyKeyListener implements KeyListener {
             Main.camera.getCoo().setY(Main.camera.getCoo().getY()-1);
         } else if (KeyEvent.getKeyText(e.getKeyCode()).equals("Q")) {
             Camera cam = new Camera(Main.camera);
-            cam.setOrientationW((float) (cam.getOrientationW()-Math.PI/2));
-            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector((float) Math.sin(cam.getOrientationW())/4, 0, (float) Math.cos(cam.getOrientationW())/4)));
+            cam.setOrientationW(cam.getOrientationW()-Math.PI/2);
+            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector(Math.sin(cam.getOrientationW())/4, 0, Math.cos(cam.getOrientationW())/4)));
         } else if (KeyEvent.getKeyText(e.getKeyCode()).equals("D")) {
             Camera cam = new Camera(Main.camera);
-            cam.setOrientationW((float) (cam.getOrientationW()+Math.PI/2));
-            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector((float) Math.sin(cam.getOrientationW())/4, 0, (float) Math.cos(cam.getOrientationW())/4)));
+            cam.setOrientationW(cam.getOrientationW()+Math.PI/2);
+            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector(Math.sin(cam.getOrientationW())/4, 0, Math.cos(cam.getOrientationW())/4)));
         } else if (KeyEvent.getKeyText(e.getKeyCode()).equals("Z")) {
-            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector((float) Math.sin(Main.camera.getOrientationW())/4, 0, (float) Math.cos(Main.camera.getOrientationW())/4)));
+            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector(Math.sin(Main.camera.getOrientationW())/4, 0, Math.cos(Main.camera.getOrientationW())/4)));
         } else if (KeyEvent.getKeyText(e.getKeyCode()).equals("S")) {
-            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector((float) -Math.sin(Main.camera.getOrientationW())/4, 0, (float) -Math.cos(Main.camera.getOrientationW())/4)));
+            Main.camera.setCoo(Vector.add(Main.camera.getCoo(), new Vector(-Math.sin(Main.camera.getOrientationW())/4, 0, -Math.cos(Main.camera.getOrientationW())/4)));
         }
 
 
         else if (KeyEvent.getKeyText(e.getKeyCode()).equals("Right")) {
-            Main.camera.setOrientationW((float) (Main.camera.getOrientationW()+0.1));
+            Main.camera.setOrientationW(Main.camera.getOrientationW()+0.1);
         } else if (KeyEvent.getKeyText(e.getKeyCode()).equals("Left")) {
-            Main.camera.setOrientationW((float) (Main.camera.getOrientationW()-0.1));
+            Main.camera.setOrientationW(Main.camera.getOrientationW()-0.1);
         } else if (KeyEvent.getKeyText(e.getKeyCode()).equals("Up")) {
-            Main.camera.setOrientationH((float) (Main.camera.getOrientationH()-0.1));
+            Main.camera.setOrientationH(Main.camera.getOrientationH()-0.1);
             if (Main.camera.getOrientationH()<0) {
                 Main.camera.setOrientationH(0);
             }
         } else if (KeyEvent.getKeyText(e.getKeyCode()).equals("Down")) {
-            Main.camera.setOrientationH((float) (Main.camera.getOrientationH()+0.1));
+            Main.camera.setOrientationH(Main.camera.getOrientationH()+0.1);
             if (Main.camera.getOrientationH()>Math.PI) {
-                Main.camera.setOrientationH((float) Math.PI);
+                Main.camera.setOrientationH(Math.PI);
             }
         }
     }
